+++ core/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add `FieldDeserializer::try_id()`.
+
 ### Changed
 
 - Bump protocol version to 16.