+++ core/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `FieldDeserializer::try_id()`.
+- Add `EnumDeserializer::try_variant()`.
 
 ### Changed